# REST-api/openapi.et.yaml
openapi: 3.0.0
info:
  title: Google Forms klooni API
  version: "1.0.0"
  description: >
    Google Forms sarnase teenuse OpenAPI näidisspetsifikatsioon.
    Kasutab ressursside nimedena üksikuid mitmuse nimisõnu ja /sessions autentimiseks.
servers:
  - url: /
    description: Tootmisserver
...
paths:
...
  /forms:
    get:
      tags:
        - forms
      summary: Kuva kõik vormid
      operationId: listForms
      responses:
        "200":
          description: Vormide nimekiri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Form"
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    post:
      tags:
        - forms
      summary: Loo uus vorm
      operationId: createForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormCreate"
      responses:
        "201":
          description: Vorm on loodud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
  /forms/{formId}:
    get:
      tags:
        - forms
      summary: Kuva konkreetse vormi üksikasjad
      operationId: getForm
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vormi üksikasjad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
        "404":
          description: Vormi ei leitud
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    patch:
      tags:
        - forms
      summary: Uuenda olemasoleva vormi osa
      operationId: updateForm
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormUpdate"
      responses:
        "200":
          description: Vorm on uuendatud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
        "404":
          description: Vormi ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    delete:
      tags:
        - forms
      summary: Kustuta vorm
      operationId: deleteForm
      parameters:
        - name: formId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Vorm on kustutatud
        "404":
          description: Vormi ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
...
  /users:
    get:
      tags:
        - users
      summary: Kuva kõik kasutajad
      operationId: listUsers
      responses:
        "200":
          description: Kasutajate nimekiri
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]

    post:
      tags:
        - users
      summary: Loo uus kasutaja (registreeru)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Kasutaja on loodud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Päringus on puudu vajalikud väljad
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Päringus on puudu vajalikud väljad: email, parool ja nimi on nõutud. Parool peab olema vähemalt 8 tähemärki pikk, Parool peab sisaldama vähemalt 3 järgnevast: suurtähed, väiketähed, numbrid, erimärgid"
        "409":
            description: Selle emailiga kasutaja on juba olemas
        "500":
          $ref: "#/components/responses/ServerError"
  /users/{userId}:
    get:
      tags:
        - users
      summary: Kuva konkreetse kasutaja üksikasjad
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Kasutaja üksikasjad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Kasutajat ei leitud
        "401":
          description: Autoriseerimata
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
    patch:
      tags:
        - users
      summary: Uuenda olemasoleva kasutaja osa
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Kasutaja on uuendatud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Kasutajat ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
    delete:
      tags:
        - users
      summary: Kustuta kasutaja
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Kasutaja on kustutatud
        "404":
          description: Kasutajat ei leitud
        "401":
          description: Autoriseerimata
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
      security:
        - BearerAuth: [ ]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Vigane päring - vigased sisendparameetrid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Serveri sisemine viga
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - message
...
    FormCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
    FormUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
    Form:
      type: object
      properties:
        id:
          type: integer
          description: Vormi unikaalne identifikaator
        userId:
          type: integer
          description: Vormi loonud kasutaja ID
        title:
          type: string
          description: Vormi pealkiri
        description:
          type: string
          description: Vormi kirjeldus
        createdAt:
          type: string
          format: date-time
          description: Loomise ajatempel
        updatedAt:
          type: string
          format: date-time
          description: Viimase uuenduse ajatempel
      required:
        - id
        - userId
        - title
        - description
        - createdAt
        - updatedAt
...
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
      required:
        - email
        - password
        - name
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        passwordUpdated:
          type: boolean
          description: Kas parool on uuendatud pärast viimast sisselogimist
      required:
        - id
        - email
        - name
        - createdAt
        - updatedAt
        - passwordUpdated